+++ taxel-gui/Cargo.toml
@@ -17,3 +17,4 @@ anyhow = { workspace = true }
 log = { workspace = true }
 quick-xml = { workspace = true }
 eframe = { workspace = true }
+dioxus-devtools = { workspace = true }
